requirements.txt
Flask==3.0.3
line-bot-sdk==3.13.0
protobuf~=5.28.2
google-api-python-client~=2.149.0